TP-205 wip(ItemNotaFiscal) criado dto, convertor, controller

+++ crud/src/main/java/com/agrotis/trainees/crud/repository/ItemNotaFiscalRepository.java
@@ -3,6 +3,7 @@ package com.agrotis.trainees.crud.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.agrotis.trainees.crud.entity.ItemNotaFiscal;
@@ -15,11 +16,11 @@ public interface ItemNotaFiscalRepository extends JpaRepository<ItemNotaFiscal, 
 
     List<ItemNotaFiscal> findByNotaFiscal(NotaFiscal notaFiscal);
 
-    List<ItemNotaFiscal> findByQuantidade(double quantidade);
+    List<ItemNotaFiscal> findByQuantidade(BigDecimal quantidade);
 
-    List<ItemNotaFiscal> findByPrecoUnitario(double precoUnitario);
+    List<ItemNotaFiscal> findByPrecoUnitario(BigDecimal precoUnitario);
 
-    List<ItemNotaFiscal> findByValorTotal(double valorTotal);
+    List<ItemNotaFiscal> findByValorTotal(BigDecimal valorTotal);
 
     List<ItemNotaFiscal> findByNotaFiscalAndProduto(NotaFiscal notaFiscal, Produto produto);
 }
